requirements.txt
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.5
google-cloud-vision==3.9.0
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
playwright==1.49.1
hypercorn==0.17.3
google-cloud-storage==2.19.0
pydantic==2.10.6
pypdf==5.3.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
python-multipart==0.0.20
pinecone==6.0.1
openai==1.77.0
alembic==1.15.2
pillow==11.2.1
